+++ index.html
@@ -42,11 +42,10 @@
           <li class="dropdown">
             <a href="javascript:void(0)" class="dropbtn">File</a>
             <div class="dropdown-content">
-              <a id="menubar-new">New file<span class="hotkey"></span></a>
-              <a id="menubar-load">Load file...<span class="hotkey"></span></a>
-              <a id="menubar-save">Save file...</a>
-              <hr />
-              <a id="menubar-monitor">VICE Snapshot Monitor</a>
+              <a id="menubar-new">New sprites<span class="hotkey"></span></a>
+              <a id="menubar-load">Load sprite data...<span class="hotkey"></span></a>
+              <a id="menubar-save">Save sprite data...</a>
+              <a id="menubar-export">Export sprite data...</a>
             </div>
           </li>
 
@@ -89,7 +88,9 @@
 
           <li class="dropdown">
             <a href="javascript:void(0)" class="dropbtn">View</a>
-            <div class="dropdown-content">
+            <div class="dropdown-content">    
+              <a id="menubar-monitor">VICE Snapshot Monitor</a>
+              <hr />
               <a id="menubar-fullscreen">Toggle Fullscreen<span class="hotkey">shift + f</span></a>
               <hr />
               <span class="label">Editor Window</span>
